Allow setting the data dir file mode
To help increase security of the allocations directory, it is sometimes
necessary to reduce the permissions of the nomad data directory such that
only the nomad daemon can access it. This change allows the administrator
to set a custom permissions on the data dir directory managed by this module.

Backward compatibility is preserved, as the default mode is still `0755`

Fixes #68